Reworked handling files with only removed lines
With the approach of explicitly specifying only the files with added lines, if we have only the ones with removed lines, the list becomes empty, and `git diff` includes all files.

Instead, use `awk` as a replacement for `grep` and `cut` which besides filtering can also alter every line of the output.

Fixes #4.