Fixed failures on macOS
The BSD `xargs` implies `--no-run-if-empty` by default and it cannot be disabled. This causes the diff to be empty in case if there's nothing to exclude.

The solution is to take the constant '.' argument from the last `git diff` and pass it to `xargs` together with the files to be excluded, so that the `xargs`s input is always non-empty.